split run client
