+++ packages/web/src/App.tsx
@@ -107,185 +107,8 @@ import { generateId } from "../../core/src/id"
 import { prettyCost, prettyDuration, prettyTokens } from "../../core/src/pretty"
 import { useEventListener } from "./useEventListener"
 import { useUrlSearchParams } from "./useUrlSearchParam"
-
-const fetchScripts = async (): Promise<Project> => {
-    const res = await fetch(`${base}/api/scripts`, {
-        headers: {
-            Accept: "application/json",
-            Authorization: apiKey,
-        },
-    })
-    if (!res.ok) throw new Error(await res.json())
-
-    const j: PromptScriptListResponse = await res.json()
-    return j.project
-}
-const fetchEnv = async (): Promise<ServerEnvResponse> => {
-    const res = await fetch(`${base}/api/env`, {
-        headers: {
-            Accept: "application/json",
-            Authorization: apiKey,
-        },
-    })
-    if (!res.ok) throw new Error(await res.json())
-
-    const j: ServerEnvResponse = await res.json()
-    return j
-}
-const fetchRuns = async (): Promise<RunResultListResponse> => {
-    const res = await fetch(`${base}/api/runs`, {
-        headers: {
-            Accept: "application/json",
-            Authorization: apiKey,
-        },
-    })
-    if (!res.ok) throw new Error(await res.json())
-
-    const j: RunResultListResponse = await res.json()
-    return j
-}
-const fetchModels = async (): Promise<ChatModels> => {
-    const res = await fetch(`${base}/v1/models`, {
-        headers: {
-            Accept: "application/json",
-            Authorization: apiKey,
-        },
-    })
-    if (!res.ok) throw new Error(await res.json())
-
-    const j: ChatModels = await res.json()
-    return j
-}
-const fetchRun = async (
-    runId: string
-): Promise<PromptScriptEndResponseEvent> => {
-    if (!runId) return undefined
-    const res = await fetch(`${base}/api/runs/${runId}`, {
-        headers: {
-            Accept: "application/json",
-            Authorization: apiKey,
-        },
-    })
-    if (!res.ok) throw new Error(await res.json())
-
-    const j: PromptScriptEndResponseEvent = await res.json()
-    return j
-}
-
-class RunClient extends WebSocketClient {
-    static readonly SCRIPT_START_EVENT = "scriptStart"
-    static readonly SCRIPT_END_EVENT = "scriptEnd"
-    static readonly PROGRESS_EVENT = "progress"
-    static readonly RUN_EVENT = "run"
-
-    runId: string
-    trace: string = ""
-    output: string = ""
-    reasoning: string = ""
-    result: Partial<GenerationResult> = undefined
-
-    private progressEventThrottled = throttle(
-        this.dispatchProgressEvent.bind(this),
-        2000,
-        {
-            edges: ["leading"],
-        }
-    )
-
-    private stderr = ""
-    constructor(url: string) {
-        super(url)
-        this.addEventListener(QUEUE_SCRIPT_START, () => {
-            this.updateRunId({ runId: "" })
-        })
-        this.stderr = ""
-        this.addEventListener(
-            MESSAGE,
-            async (ev) => {
-                const data = (ev as MessageEvent<any>).data as
-                    | PromptScriptResponseEvents
-                    | RequestMessages
-                    | LogMessageEvent
-                switch (data.type) {
-                    case "log": {
-                        const fn = console[data.level]
-                        fn?.(data.message)
-                        break
-                    }
-                    case "script.progress": {
-                        this.updateRunId(data)
-                        if (data.trace) this.trace += data.trace
-                        if (data.output && !data.inner) {
-                            this.output += data.output
-                        }
-                        if (data.reasoning) this.reasoning += data.reasoning
-                        if (data.responseChunk) {
-                            this.stderr += data.responseChunk
-                            const lines = this.stderr.split("\n")
-                            for (const line of lines.slice(0, lines.length - 1))
-                                console.debug(line)
-                            this.stderr = lines.at(-1)
-                        }
-                        this.progressEventThrottled()
-                        break
-                    }
-                    case "script.end": {
-                        this.updateRunId(data)
-                        if (data.result) {
-                            this.result = cleanedClone(data.result)
-                        } else {
-                            const { result, trace } =
-                                (await fetchRun(data.runId)) || {}
-                            this.result = cleanedClone(result)
-                            this.trace = trace || ""
-                        }
-                        this.output = this.result?.text || ""
-                        this.reasoning = this.result?.reasoning || ""
-                        this.dispatchEvent(
-                            new CustomEvent(RunClient.SCRIPT_END_EVENT, {
-                                detail: this.result,
-                            })
-                        )
-                        this.dispatchProgressEvent()
-                        break
-                    }
-                    case "script.start":
-                        this.updateRunId(
-                            data.response as PromptScriptStartResponse
-                        )
-                        this.dispatchEvent(
-                            new CustomEvent(RunClient.SCRIPT_START_EVENT, {
-                                detail: data.response,
-                            })
-                        )
-                        break
-                    default: {
-                        console.log(data)
-                    }
-                }
-            },
-            false
-        )
-    }
-
-    private dispatchProgressEvent() {
-        this.dispatchEvent(new Event(RunClient.PROGRESS_EVENT))
-    }
-
-    private updateRunId(data: { runId: string }) {
-        const { runId } = data
-        if (runId !== this.runId) {
-            this.runId = runId
-            if (this.runId) {
-                this.trace = ""
-                this.output = ""
-                this.result = undefined
-                this.stderr = ""
-            }
-            this.dispatchEvent(new Event(RunClient.RUN_EVENT))
-        }
-    }
-}
+import { RunClient } from "./RunClient"
+import { fetchEnv, fetchScripts, fetchRuns, fetchModels, fetchRun } from "./api"
 
 type ImportedFile = FileWithPath & { selected?: boolean }
